Introduce cuda_driver_library_available(cuda_version_two_digits)
This returns "none", "compat" or "native".

"native": If the system library is new enough
"compat": If the compat libraries can be used via LD_LIBRARY_PATH
(needs restricted and at least 10.2 driver)
"none": Anything else: this CUDA can't be used

A check for "compat" can then be used in cudacore modules to load
the cuda-compat module, if needed.