Stricter checking for CUDA compat driver compatibility
New Feature Branches (such as 495.xx) are not supported targets for CUDA Forward Compatibility.
Other drivers (PB and LTSB) have EOL dates and can't be used with CUDA
versions released after those dates.

So far we've only checked this with LTSB 418 and CUDA 11.6 but using a
table we can add a max CUDA for all branches. In particular this makes
CUDA 12.2 invisible on Cedar with the 530 driver, as this NFB isn't
supported.

Tested with a few (forced) combinations on build-node.

See https://docs.nvidia.com/deploy/cuda-compatibility/ and
https://docs.nvidia.com/datacenter/tesla/drivers/index.html